+++ kll.hpp
@@ -15,15 +15,15 @@ class KLL{
         bool murmurHashSample(long element); // indica si el elemento es seleccionado al samplear
         bool reservoirKLLSample(long element, long elementWeight);
 
-        void add(long element); // agregar element al sketch
-        void addv(long element); // agregar element al sketch
-        unsigned long rank(long element); // indica el rank del elemento proporcionado
+        void add(double element); // agregar element al sketch
+        void addv(double element); // agregar element al sketch
+        unsigned long rank(double element); // indica el rank del elemento proporcionado
         long select(long rank); // retorna el elemento cuyo rank es el indicado
         long quantile(double q); // retorna elemento encontrado en el quantil q
 
         // Operaciones para realizar merge
         long height();
-        pair<vector<long>, long> sketchAtLevel(long nivel);
+        pair<vector<double>, long> sketchAtLevel(long nivel);
         bool sortedAtLevel(long nivel);
         void update(KLL kll2);
         void setSeconds(vector<long> seconds);
@@ -35,20 +35,20 @@ class KLL{
         KLL kllMerge(KLL &kll2);
 
     private:
-        // reservoirKLLSample
+        // variables para reservoirKLLSample
         unsigned long sampleWeight;
-        long sampleElement;
+        double sampleElement;
 
         unsigned int numArreglos;
         unsigned int numArreglosOcupados;
-        vector<pair<vector<long>, long> > sketch; // arreglo de arreglos con tamaño decreciente
+        vector<pair<vector<double>, long> > sketch; // arreglo de arreglos con tamaño decreciente
             // sketch[i].first almacena los vectores donde se almacenan los elementos de nivel i
             // sketch[i].second mantiene el num de elementos ocupados en dicho nivel i
         vector<bool> sorted; // indica si el sketch[i] se encuentra ordenado
         
         // Operaciones
-        void insertElement(long nivel,long &element);
-        void insertCompactionElement(long nivel,long &element,bool updating);
+        void insertElement(long nivel,double &element);
+        void insertCompactionElement(long nivel,double &element,bool updating);
         void compaction(long nivel,bool updating);
 
         // variables k y c son ctes. entregadas por el usuario, c esta en rango ]0.5,1[
